requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
base58==1.0.3
bitarray==2.7.3
cached-property==1.5.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
commonmark==0.9.1
cryptography==40.0.2
cytoolz==0.12.1
dataclassy==0.11.1
decorator==5.1.1
Deprecated==1.2.13
eip712==0.2.1
eth-abi==4.0.0
eth-account==0.8.0
eth-ape==0.6.8
eth-bloom==2.0.0
eth-hash==0.5.1
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-tester==0.8.0b3
eth-typing==3.3.0
eth-utils==2.1.0
ethpm-types==0.4.4
evm-trace==0.1.0a18
exceptiongroup==1.1.1
executing==1.2.0
frozenlist==1.3.3
greenlet==2.0.2
hexbytes==0.3.0
idna==3.4
ijson==3.2.0.post0
importlib-metadata==6.6.0
iniconfig==2.0.0
ipython==8.13.1
jedi==0.18.2
jsonschema==4.17.3
lru-dict==1.1.8
matplotlib-inline==0.1.6
morphys==1.0
msgspec==0.14.2
multidict==6.0.4
mypy-extensions==0.4.4
numpy==1.24.3
packaging==23.1
pandas==1.5.3
parsimonious==0.9.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.0.0
prompt-toolkit==3.0.38
protobuf==4.22.3
ptyprocess==0.7.0
pure-eval==0.2.2
py-cid==0.3.0
py-ecc==6.0.0
py-evm==0.6.1a2
py-geth==3.12.0
py-multibase==1.0.3
py-multicodec==0.2.1
py-multihash==0.2.3
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.7
pyethash==0.1.27
PyGithub==1.58.1
Pygments==2.15.1
PyJWT==2.6.0
PyNaCl==1.5.0
pyrsistent==0.19.3
pytest==7.3.1
python-baseconv==1.2.2
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.29.0
rich==12.6.0
rlp==3.0.0
safe-pysha3==1.0.3
semantic-version==2.10.0
six==1.16.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.12
stack-data==0.6.2
tomli==2.0.1
toolz==0.12.0
tqdm==4.65.0
traitlets==5.9.0
trie==2.1.0
typing_extensions==4.5.0
urllib3==1.26.15
varint==1.0.2
watchdog==2.3.1
wcwidth==0.2.6
web3==6.2.0
websockets==11.0.2
wrapt==1.15.0
yarl==1.9.2
zipp==3.15.0
